+++ scrapy_poet/utils/testing.py
@@ -12,6 +12,7 @@ from scrapy.settings import Settings
 from scrapy.utils.defer import maybe_deferred_to_future
 from scrapy.utils.python import to_bytes
 from scrapy.utils.test import get_crawler as _get_crawler
+from twisted.internet.defer import inlineCallbacks
 from twisted.internet.task import deferLater
 from twisted.web.resource import Resource
 from twisted.web.server import NOT_DONE_YET
@@ -121,7 +122,26 @@ class ProductHtml(HtmlResource):
     """
 
 
-async def crawl_items(
+@inlineCallbacks
+def crawl_items(spider_cls, resource_cls, settings, spider_kwargs=None, port=None):
+    """Use spider_cls to crawl resource_cls. URL of the resource is passed
+    to the spider as ``url`` argument.
+    Return ``(items, resource_url, crawler)`` tuple.
+    """
+    warn(
+        "crawl_items is deprecated; use crawl_items_async instead",
+        DeprecationWarning,
+        stacklevel=2,
+    )
+    spider_kwargs = {} if spider_kwargs is None else spider_kwargs
+    crawler = make_crawler(spider_cls, settings)
+    with MockServer(resource_cls, port=port) as s:
+        root_url = s.root_url
+        yield crawler.crawl(url=root_url, **spider_kwargs)
+    return crawler.spider.collected_items, s.root_url, crawler
+
+
+async def crawl_items_async(
     spider_cls, resource_cls, settings, spider_kwargs=None, port=None
 ):
     """Use spider_cls to crawl resource_cls. URL of the resource is passed
@@ -136,13 +156,38 @@ async def crawl_items(
     return crawler.spider.collected_items, s.root_url, crawler
 
 
-async def crawl_single_item(
+@inlineCallbacks
+def crawl_single_item(
     spider_cls, resource_cls, settings, spider_kwargs=None, port=None
 ):
     """Run a spider where a single item is expected. Use in combination with
     ``capture_exceptions`` and ``CollectorPipeline``
     """
-    items, url, crawler = await crawl_items(
+    warn(
+        "crawl_single_item is deprecated; use crawl_single_item_async instead",
+        DeprecationWarning,
+        stacklevel=2,
+    )
+    items, url, crawler = yield crawl_items(
+        spider_cls, resource_cls, settings, spider_kwargs=spider_kwargs, port=port
+    )
+    try:
+        item = items[0]
+    except IndexError:
+        return None, url, crawler
+
+    if isinstance(item, dict) and "exception" in item:
+        raise item["exception"]
+    return item, url, crawler
+
+
+async def crawl_single_item_async(
+    spider_cls, resource_cls, settings, spider_kwargs=None, port=None
+):
+    """Run a spider where a single item is expected. Use in combination with
+    ``capture_exceptions`` and ``CollectorPipeline``
+    """
+    items, url, crawler = await crawl_items_async(
         spider_cls, resource_cls, settings, spider_kwargs=spider_kwargs, port=port
     )
     try: